Title: Created a basic TPC Client/Server Program with BSD Socket API
Description:

Created a simple TCP Server that listens for incoming connections, accepts client connection, receives messages
and displays them. The code initializes a socket, binds it to a specified IP Address and Port, listens for
connections and accepts incoming connections from client. Received messages are displayed until the client sends
'quit' or closes the connection.

Created a simple TCP Client that connects to a server, sends messages, and terminates upon user input. The code
initializes a socket, establishes a connection with the server (using either a specified IP Address or DNS lookup
for host name resolution) and sends user input messages. The client prompts for messages and sends the messages to
the server until 'quit' is entered.